requirements.txt
streamlit==1.26.0
tensorflow==2.13.0
numpy==1.24.3
Pillow<=10.0.0
matplotlib==3.7.2